Support escaped commas in service tags for connect (#1532)
* support escaped commas in service tag annotations
