+++ control-plane/catalog/to-consul/resource_test.go
@@ -1490,50 +1490,6 @@ func TestServiceResource_MirroredPrefixNamespace(t *testing.T) {
 	})
 }
 
-func TestParseTags(t *testing.T) {
-	cases := []struct {
-		tagsAnno string
-		exp      []string
-	}{
-		{
-			"tag",
-			[]string{"tag"},
-		},
-		{
-			",,removes,,empty,elems,,",
-			[]string{"removes", "empty", "elems"},
-		},
-		{
-			"removes , white  ,space ",
-			[]string{"removes", "white", "space"},
-		},
-		{
-			`\,leading,comma`,
-			[]string{",leading", "comma"},
-		},
-		{
-			`trailing,comma\,`,
-			[]string{"trailing", "comma,"},
-		},
-		{
-			`mid\,dle,com\,ma`,
-			[]string{"mid,dle", "com,ma"},
-		},
-		{
-			`\,\,multi\,\,,\,com\,\,ma`,
-			[]string{",,multi,,", ",com,,ma"},
-		},
-		{
-			`  every\,\,   ,  thing  `,
-			[]string{"every,,", "thing"},
-		},
-	}
-
-	for _, c := range cases {
-		require.Equal(t, c.exp, parseTags(c.tagsAnno))
-	}
-}
-
 // lbService returns a Kubernetes service of type LoadBalancer.
 func lbService(name, namespace, lbIP string) *apiv1.Service {
 	return &apiv1.Service{
